🧹 fix: clearme command with discord api limit

+++ commands/hacker.js
@@ -109,6 +109,9 @@ async function clearme (client, message, process) {
 
   if (!args[1] || isNaN(parseInt(args[1]))) { return message.channel.send('❌ Please indicate a number of messages to delete.') }
 
+  if (parseInt(args[1]) > 100) { return message.channel.send('❌ You can only delete up to 100 messages at a time.') }
+  if (parseInt(args[1]) < 1) { return message.channel.send('❌ You must delete at least 1 message.') }
+
   await (await message.channel.messages.fetch({ limit: (parseInt(args[1]) + 1) })).filter((msg) => msg.author.id === client.user.id).forEach(m => m.delete())
   const allIsDelete = await message.channel.send(`✅ Deletion of ${args[1]} messages.`)
   await setTimeout(() => { allIsDelete.delete() }, 2500)
